moved the applyUnitsOptions to the datasets-filter.component, small adjustments
